Add support for in memory caching
Introduces an alternate to the legacyCollections and
genericCollection that relies on caching resource in memory. The
new collection[T] takes inspiration from the legacy collection types
but aims to both simplify and break reliance on using the local
storage services.

The two collection mechanisms will and can coexist until all types
are converted to the new mechanism. This includes an initial
conversion of static tokens, users, and cert authorities to prove
out the effectivenss of the strategy.

Additionally, during the conversion process some refactoring will
be done to try to better organize and reduce the size of cache.go.
In this PR all of the code specific to static tokens was moved
out of cache.go and legacy_collections.go and into static_tokens.go.
This will allow separating concerns and making it much easier to
identify where specific content lives.
