+++ api/types/user.go
@@ -156,6 +156,8 @@ type User interface {
 	SetWeakestDevice(MFADeviceKind)
 	// GetWeakestDevice gets the MFA state for the user.
 	GetWeakestDevice() MFADeviceKind
+	// Clone creats a copy of the user.
+	Clone() User
 }
 
 // NewUser creates new empty user
@@ -271,14 +273,15 @@ func (u *UserV2) SetName(e string) {
 	u.Metadata.Name = e
 }
 
+func (u *UserV2) Clone() User {
+	return utils.CloneProtoMsg(u)
+}
+
 // WithoutSecrets returns an instance of resource without secrets.
 func (u *UserV2) WithoutSecrets() Resource {
-	if u.Spec.LocalAuth == nil {
-		return u
-	}
-	u2 := *u
+	u2 := utils.CloneProtoMsg(u)
 	u2.Spec.LocalAuth = nil
-	return &u2
+	return u2
 }
 
 // GetTraits gets the trait map for this user used to populate role variables.
