+++ lib/cache/resource_reverse_tunnel.go
@@ -30,7 +30,7 @@ func (c *Cache) ListReverseTunnels(ctx context.Context, pageSize int, pageToken 
 	ctx, span := c.Tracer.Start(ctx, "cache/ListReverseTunnels")
 	defer span.End()
 
-	rg, err := readCollectionCache(c, c.collections.reverseTunnels)
+	rg, err := readLegacyCollectionCache(c, c.legacyCacheCollections.reverseTunnels)
 	if err != nil {
 		return nil, "", trace.Wrap(err)
 	}
